Satisfy clippy in pendantic mode.

+++ src/web.rs
@@ -17,9 +17,9 @@ pub async fn serve() {
     dotenv().ok();
     let address = env::var("SERVER_ADDRESS").expect("SERVER_ADDRESS must be set");
     let port = env::var("SERVER_PORT").expect("SERVER_PORT must be set");
-    let socket_address: SocketAddr = format!("{}:{}", address, port).parse().expect("Unable to create a valid socket address.");
+    let socket_address: SocketAddr = format!("{address}:{port}").parse().expect("Unable to create a valid socket address.");
 
-    let app = get_app();
+    let app = async { get_app() };
     axum::Server::bind(&socket_address)
         .serve(app.await.into_make_service())
         .await
@@ -27,7 +27,7 @@ pub async fn serve() {
 }
 
 // Returns the app routes.
-async fn get_app() -> Router {
+fn get_app() -> Router {
     Router::new()
         .route("/", get(index))
         .route("/adjustment-types", get(list_adjustment_types))
@@ -40,7 +40,7 @@ async fn get_app() -> Router {
 // Handler for the main API endpoint. Returns the version of the API as a JSON object.
 async fn index() -> impl IntoResponse {
     let version = env!("CARGO_PKG_VERSION");
-    let response = Response::new(Body::from(format!("{{\"version\": \"{}\"}}", version)));
+    let response = Response::new(Body::from(format!("{{\"version\": \"{version}\"}}")));
     (StatusCode::OK, response)
 }
 
@@ -54,16 +54,13 @@ async fn list_adjustment_types() -> impl IntoResponse {
 // GET handler: shows the adjustment type with the given ID.
 async fn get_adjustment_type(Path(id): Path<u64>) -> impl IntoResponse {
     let adjustment_type = db::get_adjustment_type(id);
-    // Return a 404 if the adjustment type does not exist.
-    match adjustment_type {
-        Some(adjustment_type) => {
-            let response = Response::new(Body::from(serde_json::to_string(&adjustment_type).unwrap()));
-            (StatusCode::OK, response)
-        }
-        None => {
-            let response = Response::new(Body::from(format!("{{\"error\": \"Adjustment type with ID {} not found\"}}", id)));
-            (StatusCode::NOT_FOUND, response)
-        }
+
+    if let Some(adjustment_type) = adjustment_type {
+        let response = Response::new(Body::from(serde_json::to_string(&adjustment_type).unwrap()));
+        (StatusCode::OK, response)
+    } else {
+        let response = Response::new(Body::from(format!("{{\"error\": \"Adjustment type with ID {id} not found\"}}")));
+        (StatusCode::NOT_FOUND, response)
     }
 }
 
@@ -71,7 +68,7 @@ async fn get_adjustment_type(Path(id): Path<u64>) -> impl IntoResponse {
 async fn create_adjustment_type(Json(payload): Json<NewAdjustmentType>) -> impl IntoResponse {
     let rows_inserted = db::add_adjustment_type(payload.description, payload.adjustment);
     // Respond with the number of inserted rows.
-    let response = Response::new(Body::from(format!("{{\"inserted\": \"{}\"}}", rows_inserted)));
+    let response = Response::new(Body::from(format!("{{\"inserted\": \"{rows_inserted}\"}}")));
     (StatusCode::CREATED, response)
 }
 
@@ -79,23 +76,21 @@ async fn create_adjustment_type(Json(payload): Json<NewAdjustmentType>) -> impl 
 async fn delete_adjustment_type(Path(id): Path<u64>) -> impl IntoResponse {
     // Return a 404 if the adjustment type does not exist.
     let adjustment_type = db::get_adjustment_type(id);
-    match adjustment_type {
-        None => {
-            let response = Response::new(Body::from(format!("{{\"error\": \"Adjustment type with ID {} not found\"}}", id)));
-            return (StatusCode::NOT_FOUND, response);
-        }
-        _ => {}
+    if adjustment_type.is_none() {
+        let response = Response::new(Body::from(format!("{{\"error\": \"Adjustment type with ID {id} not found\"}}")));
+        return (StatusCode::NOT_FOUND, response);
     }
+
     let result = db::delete_adjustment_type(id);
     match result {
         Ok(rows_deleted) => {
             // Respond with the number of deleted rows.
-            let response = Response::new(Body::from(format!("{{\"deleted\": \"{}\"}}", rows_deleted)));
+            let response = Response::new(Body::from(format!("{{\"deleted\": \"{rows_deleted}\"}}")));
             (StatusCode::OK, response)
         }
         Err(e) => {
             // Respond with an error message.
-            let response = Response::new(Body::from(format!("{{\"error\": \"{}\"}}", e)));
+            let response = Response::new(Body::from(format!("{{\"error\": \"{e}\"}}")));
             (StatusCode::BAD_REQUEST, response)
         }
     }
@@ -103,7 +98,7 @@ async fn delete_adjustment_type(Path(id): Path<u64>) -> impl IntoResponse {
 
 // GET handler: lists the available adjustments, optionally filtered by adjustment type and limit.
 async fn list_adjustments(Query(filter): Query<db::AdjustmentQueryFilter>) -> impl IntoResponse {
-    let adjustments = db::get_adjustments(filter);
+    let adjustments = db::get_adjustments(&filter);
     let response = Response::new(Body::from(serde_json::to_string(&adjustments).unwrap()));
     (StatusCode::OK, response)
 }
